proxy tests
