+++ packages/gitbook/package.json
@@ -10,6 +10,7 @@
         "typecheck": "tsc --noEmit",
         "e2e": "playwright test e2e/internal.spec.ts",
         "e2e-customers": "playwright test e2e/customers.spec.ts",
+        "e2e-proxy": "playwright test e2e/proxy.spec.ts",
         "unit": "bun test {src,packages}",
         "generate": "gitbook-icons ./public/~gitbook/static/icons custom-icons && gitbook-math ./public/~gitbook/static/math",
         "copy:icons": "gitbook-icons ./public/~gitbook/static/icons",
